EVG-6517 Support generate.tasks across deploys
Prior to this commit the retries for generate.tasks are in the body of the job
itself. This means that canceling the job's context would cause it to error.
This commit moves the retry logic into amboy by requeueing the job with an
incrementing ID. This ID is stored in the task document to support the poll job.
The poll job must also check for legacy task IDs, since people may restart old
tasks.

I also snuck in extending the polling interval for jobs.