EVG-6517 Do not mark task as having generated if it raced
In the case of a race generate.tasks returns document not found, since the
document had been modified by another job. This commit fixes a case where we
failed to check this. In this case we should not set the GeneratedTasksKey
field, since this job has not in fact generated tasks.