EVG-6517 Don't generate tasks if the generator is not running
If a task is aborted or timed out, there may still be generate jobs in the
queue. However, they should not run. In the former case, the user has indicated
that they don't want them to run by aborting the task. In the latter case, if
the generated tasks end up in a display task with the generator, then that task
will be purple or red even if all the generated tasks succeed. The user may
restart this task, causing not just the generator, but all the other tasks to
reexecute.