EVG-6918 Handle additional error case in CreateFleet
In instant mode CreateFleet can error without returning an error. This means we
have to add an additional check for that case so that we do the right thing in
the RequestLimitExceeded case. Currently we are not doing exponential back off
for CreateFleet the way we do for every other API all.