EVG-6974 Keep diffbox visible despite long test names (#2869)
When a test with a long name appears in this table, it prevents text wrapping in its table cell, which makes the table too wide for its scrolling container. Because the container has overflow-x: none, you can't scroll to see the red/green diffboxes.

The fix is to allow text wrapping to break in the middle of a word. This prevents the text from forcing the table to be wider than its container.