EVG-7226 Check if LocalFile is absolute after expanding values
Prior to this commit s3.put checked if LocalFile was an absolute path prior to
expanding values. This means that if a user provided an expansion in LocalPath
which made it absolute, s3.put would behave as if it were local. This commit
inverts the order, expanding values before checking if the path is absolute.