EVG-7681 Count provisioning host status task group task count (#3470)
* EVG-7681 Count provisioning host status task group task count

New host provisioning allows hosts to begin to run tasks before they are in the
running status. This means that new hosts might not be counted by the
NumHostsByTaskSpec query, which causes a race in task group task dispatching in
which multiple tasks in a single-host task group can run concurrently on a
single host.