EVG-8097
EVG-8097 There is a checkbox on the distro page that currently blocks new jobs from being queued, but does not dequeue existing jobs. In order to make the alerts for MCI-3261 more meaningful, this improvement will make it so that quarantining hosts will not trip this alert by producing a queue > 0 with 0 healthy hosts.