EVG-14522 Return http.Client to pool after github.Client is done (#4695)
* EVG-14522 Return http.Client to pool after github.Client is done

Before this change, Execute() called executeLoop() in a loop. executeLoop()
created a github.Client if one did not already exist in its receiver. However,
to create this github.Client, it first constructed an http.Client by asking for
one from the httpClient pool. It deferred putting it back into the pool until
the end of executeLoop(). However, this is a bug. Putting it back in the pool
unset its token. The next time the loop in Execute() calls executeLoop(), the
http.Client's token was unset.

Separately, this change increases the time between retries for checking the
mergeability of a pull request. The GitHub docs explicitly ask that you poll,
which we do. However,
https://docs.github.com/en/rest/guides/getting-started-with-the-git-database-api
does not explain what a reasonable polling interview is. I suspect that
EVG-14522 occurred because we only polled for 10 seconds, which is occasionally
not enough. This PR changes this to approximately 1 minute.