EVG-14451 Remove goroutine from streamArchiveContents
Using a goroutine to archive files does help with performance, but it cannot
cut time by more than 50%, since there is only a single producer and consumer.
The tradeoff is that with a goroutine we cannot pass the error from walking the
filesystem to the caller, causing us to occasionally panic, since the error
from filepath.Walk means we shouldn't check what it returned.