EVG-16635: Update EC2 spot instance creation to handle UnfulfillableCapacity Errors (#5725)
* EVG-16635: Pull fallback code into ec2 cloud manager.

Instead of having specialized code to handle EC2 error codes in
provisioning_create_host.go, this moves the logic into ec2.go
and handles the retry and fallback as part of making the spot
request. This makes the levels of abstraction more consistent.

This also adds handling for the Unfulfillable capacity errors that we have
been seeing from Amazon, and ensure we retry as OnDemand requests.