BUILD-17945: use env vars for mac.sign secrets
Previously, mac.sign was invoked with the API secret passed on the
command line. As per EVG-20875, it is unsafe to pass command-line secrets
to anything on a machine running the Evergreen agent because it could be
captured by the "ps" commands the agent periodically runs and stores in
system logs. This commit changes mac.sign to receive its API secret as
an environment variable. The macnotary binary already supports
configuration like this.