DEVPROD-6292: require matches for papertrail.trace (#7735)
This commit makes a small change to papertrail.trace so it returns an error if none of the filename patterns in use match a file on disk.

Additionally, it fixes a bug and adds a test associated with working directory parsing. This bug is what prompted failing without matches, because it was very hard to debug without the error. The bug was fixed by switching to GetWorkingDirectory().