DEVPROD-6269: skip papertrail.trace testing on mac
This commit skips integration testing papertrail.trace on MacOS because
all MacOS hosts currently run outside of AWS. Hosts running outside of
AWS cannot bypass corpsecure when talking to Papertrail. This will
change in the future when we migrate MacOS hosts to AWS, but for now we
should skip testing there and document that this is unsupported.

Additionally, this commit adds documentation letting users know that
papertrail.trace is unavailable from outside of AWS.