DEVPROD-6548: set author for periodic builds
This commit ensures that the author is set when tasks are created via
a periodic build. Right now, the VersionMetadata field is initialized
with only the most recent revision we fetched from the database. This
commit changes the behavior so that we also fetch the author and populate
the metadata with the author of the same commit. All versions should have
an author associated with them.