DEVPROD-12040: Add conditional writes to s3.put
This commit updates s3.put to support S3's new conditional writes
feature. The conditional writes feature guarantees that a write will not
succeed if the remote key already exists. This check respects s3's
strong read-after-write data model.

In order to enable this feature, I updated to the latest version of pail.
This required bumping the minimum go version of this project to 1.21.

While adding this feature, I realized that the existing skip_existing
implementation is buggy and has not been working correctly because of the
final check we do for updatedFiles equaling fileList. This commit also
fixes that bug.