DEVPROD-15875: Add Optibot configuration file for AI code review (#8838)
This adds a configuration file that is used by Optibot, an AI-powered code reviewer, to decide how to operate.

I disabled auto-summarization of a PR because we already tend to do that ourselves and have a clear template for it. This can be turned on later if we get lazy. But I enabled auto-review so that we can see how well it performs.

If we don't like the quality of the code reviews, change "reviews.auto" to false and that should turn it off.

Auto-approvals are also off. I think that would make it accept a PR if it doesn't find issues. That seems premature at this point.