DEVPROD-16983: update to latest pail version (#8968)
This commit bumps our version of pail to the latest version in order to realize large performance gains with s3.put. Because commit 845d211 briefly broke task logs, I have updated one of our existing integration tests to check that task logs were uploaded correctly. This test will flush all of its logs and attempt to download them again. I have confirmed that checking out the bad commit causes the test to fail, and that it passes with the new version of pail. As far as I can tell this is the first automated test using real resources we have for the logging system. I think this will give us a lot of confidence in making changes to logging long-term.