DEVPROD-18197 Consider hosts with dependencies met in host drawdown (#9468)
When considering host drawdown, currently we check the total task queue length to figure out roughly if there's still tasks to run in the queue in case it's in between tasks and will soon pick one up. Instead, we shall now be checking the queue for tasks whose dependencies have been met. This would be an improvement as we don't want hosts to idle while waiting to run a task until the task's dependencies are met.